Problem: [/changedCurrencies] Bad UI/UX #1276
Solution: Bootswatchified the page and showing only one proposed change
at a time.
